Strona gotowa dane poprawione done

+++ backend/drug/src/main/java/HD/projekt/drug/Repozytorium/DimTimeRepository.java
@@ -3,5 +3,5 @@ package HD.projekt.drug.Repozytorium;
 import HD.projekt.drug.entity.DimTime;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DimTimeRepository extends JpaRepository<DimTime, Integer> {
+public interface DimTimeRepository extends JpaRepository<DimTime, Long> {
 }
