+++ backend/drug/src/main/java/HD/projekt/drug/Repozytorium/FactDrugRepository.java
@@ -3,5 +3,5 @@ package HD.projekt.drug.Repozytorium;
 import HD.projekt.drug.entity.FactDrug;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FactDrugRepository extends JpaRepository<FactDrug, Integer> {
+public interface FactDrugRepository extends JpaRepository<FactDrug, Long> {
 }
